% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmhc.R
\name{netmhc}
\alias{netmhc}
\title{Call netMHC for a peptide}
\usage{
netmhc(peptide, netmhc.dir, hla)
}
\arguments{
\item{peptide}{Peptide sequence}

\item{hla}{HLA type}
}
\description{
Call netMHC for a peptide
}
