% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run.netmhc}
\alias{run.netmhc}
\title{Wrapper Script for netMHCpan}
\usage{
run.netmhc(input, netmhc.dir, progress.bar = TRUE)
}
\arguments{
\item{input}{Data frame with columns \code{peptide}, \code{hla}}

\item{netmhc.dir}{netMHCpan source directory}

\item{progress.bar}{Display progress bar}
}
\description{
Wrapper Script for netMHCpan
}
