% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildtype.R
\name{wildtype}
\alias{wildtype}
\title{Finds the wild-type sequence of a peptide}
\usage{
wildtype(neoag, z = NA)
}
\arguments{
\item{neoag}{Data.frame with columns \code{peptide}, \code{gene}, and \code{mutation}}

\item{z}{List of known sequences for all genes to be considered}
}
\description{
Finds the wild-type sequence of a peptide
}
